TC/OC members, communication, fixes
Added subsection to preamble for TC and OC members, please fill in the
OC members I missed.

Added Communication subsections on remote control and monitoring,
important considerations as to keep the competition fair.

Fixed a few things, in particular escape an ampersand in the last commit.
Add section label for referencing. Please add these as you go through
the sections you work on.


git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@15 615d7b66-ede9-4446-987f-3b89ec2f13a0