Rotate coordinate system
This introduces a 90° counter-clockwise rotation of the playing field
and a reset of the origin (again) to the lower left corner.

The reason is twofold. First and foremost, we want to make it simpler
to map the rulebook layout (which will often translate into GUIs) to
the environment as it is perceived. During RoboCup and supposedly also
in many test setup typically observers and in particular referees and
machine operators sit at a side without the input storage or delivery
zone. Second, the new map layout more closely resembles (visually) a
soccer field more easily attracting RoboCupers from other leagues.

I hope I found most places. Will read again to make sure next week.


git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@28 615d7b66-ede9-4446-987f-3b89ec2f13a0