Unify to American English, spell-check, tweaks
Unify the used form of English to the American version. It was used in the
larger parts already and there are potentially more robotics people there.

Run a spell-check and fix several things.

Rename CPPS to CPSPE, we might want to shorten this to CPS in prodcution
environments, e.g. not coining a new term. Let's discuss.


git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@31 615d7b66-ede9-4446-987f-3b89ec2f13a0