Fixed (degermanification) some stuff on the first pages, contribution to the main part incoming. Please do not use so many comma as they are only used in rare cases in AE. I would strongly vote in favor of the Oxford comma, though.
If you further want to dischuss the CPS-CPPS stuff please consider this document first (german only) and then let me know if you still feel repellent towards this terminology.
http://www.bmbf.de/pubRD/Umsetzungsempfehlungen_Industrie_4.0_final_2012-10-02.pdf

As for the question to wether use (R) with Robotino: Our policy is to use $^{\textregistered}$, so ^(R)...which my poor TeX skills were unable to incorporate.

git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@37 615d7b66-ede9-4446-987f-3b89ec2f13a0