Title, fixes, added robot maintenance subsection
Changed the title to be more in line with other RoboCup rule books.
Fixes on the text here and there spread over the whole text. In particular,
moved game relevant descriptions out of field setup descriptions, i.e.
separate environment from gameplay descriptions.
The section that got considerable modifications is about interruptions
and robot maintenance. Make it more explicit and clear what can be done,
and how, when a teams wants to take out a robot. Also make clear that the
field referee has the power to decide if and when and has the last word.


git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@47 615d7b66-ede9-4446-987f-3b89ec2f13a0