Update late order rules
The late order was described in two different sections. Pull this together to
a single section and clarify a few more things. In particular make clear that
the order will request a P3 to be delivered within 120 sec (this will most
likely change next year, but for now it is fixed).


git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@57 615d7b66-ede9-4446-987f-3b89ec2f13a0