Remove inactive robot rule.
After thought and debate we came to the conclusion that this rule does more
harm than good, just consider the following three cases:

- a team indeed decides to use the old vision-triggered late order behavior.
  A robot standing next to the insertion slot might easily be deemed inactive.
- a team might place a robot close to the delivery gates to observe the active
  gate
- a team which only plays with two robots on the field might have a third
  robot on standby, and if one active field robot is taken out the robot might
  jump in to replace it -- autonomously!
  This kind of failover behavior is something we should encourage.

Additionally, there were many questions by referees at the German Open unsure
when to assuma robot inactive -- often with the argument that only the team
(or the robot) can know if something useful is actually going on.

The rule was adapted from soccer robots with two competing teams. Once we have
competing teams in the arena, we might re-consider our position and add such a
rule again.


git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@65 615d7b66-ede9-4446-987f-3b89ec2f13a0