Added robot marking section
We need markings on the robot in order to distinguish them. At the moment it
is virtually impossible for the human referee to properly track maintenance
times. Still when doing this through the refbox, there must be a clear way to
identify the robots. For Eindhoven, I will pring a couple of hundred
pre-printed labels, so there should be no problem.

Additionaly I have added the requirement of a head marker for the overhead
camera system.


git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@73 615d7b66-ede9-4446-987f-3b89ec2f13a0