Referee must take out puck delivered to inactive gate
If a puck was errorneously delivered to an inactive gate it has to be
removed. This means that a puck is removed in any case as soon as a delivery
gate has accepted it that is if the active gate showed all lights on or an
inactive get red on yellow flashing, but not if the active gate is just
yellow flashing.


git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@75 615d7b66-ede9-4446-987f-3b89ec2f13a0