Intro to MPS challenge, section headings, links
Give a short intro to what the MPS is, provide a pointer to additional
information. Make the Free Challenge and Conducting sections actual
sections so they appear on their own in the table of contents.

Re-structure the links, they are now are one level deeper on our website
under "/links/" which makes it obvious that this is not our content but
merely pointing somewhere. Use this for MPS and Robotino 3.

git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@128 615d7b66-ede9-4446-987f-3b89ec2f13a0