update the Robotino System section.
I removed the descriptions about the RFID device and the Pushind Device because we don't use them in 2015.

And the limitation of height is exceeded from 0.7m to 1.1m.
Because the height of MPS station is about 0.85m.
I think 0.25cm is enough height to mount the laptop computer or somethings else.
If need, however, please modify the maximum height.

I set the maximum size of gripper is same as the additional goods, because it is difficult for us to distinguish between the additional goods and the gripper.

For safety, the gripper must remove the workpiece if the referee wants to do it.

And the colored markers must be weared on the top and side of robotinos for audiences instead of round labels for the overhead camera system.


git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@149 615d7b66-ede9-4446-987f-3b89ec2f13a0