production: distinguish scheduled and broken machine downtime
The scheduled downtime works as before. I have reduced the number from
6-8 times to 2-4 times given the reduced number of machines and because
two colors are influenced by a single RS. We might want to tweak this
further, e.g. only CS that break.

Additionally mis-using machines, either by not setting the machine up
via communication, sending an invalid message, or simply by physical
mis-interaction will also result in a down-time. In the former cases it
is used to penalize misbehaving robots. In the latter cases it is
required for the referee to assess the situation and deal with it.

git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@155 615d7b66-ede9-4446-987f-3b89ec2f13a0