Increase points for mounting final ring
Reasoning:
The (rough) number of processing steps and points for each complexity is
as follows (assuming the "worst" requirement for additional bases and
counting in particular travel requirements as steps):

Complexity	Steps		Points		Relative Increase
C0		 8		 22		n/a
C1		14		 32 --  42	45% --  91%
C2		18		 52 --  67	24% -- 109%
C3		20		102 -- 117	52% -- 125%

The relative increase means the increase of points compared to the lower
complexity class. For example, from C3 to C2 the relative increase is
from the least beneficial upgrade, i.e. from C2 with 3 additional bases
total to C3 with 1 additional base total, that is (102-67)/67=24%, to
the best possible upgrade, i.e. from C1 with no additional base to C3
with 3 additional bases, that is (117-52)/52=125%.

So producing two C0 does score more than one C1, but the cost (number of
steps) is much higher thus making it less favorable. The same holds for
the other complexity to the next higher complexity.

These seem to be reasonable first guess for balancing risk, cost, and
reward. But obviously this is subject to change once we have some actual
data on the matter.

git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@238 615d7b66-ede9-4446-987f-3b89ec2f13a0