Update introduction with more buzzwords and 2016
Christian, as noted in my earlier hangout messages I would like to
request feedback especially regarding the sentence describing Festo as
the industrial partner of this league.
We have spoken about this informally on several occasions. Let me know
if this is an adequate summary of these conversations of Festo's role.

git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@271 615d7b66-ede9-4446-987f-3b89ec2f13a0