Clarifications in machine coordinate system
That is a rather complicated explanation that would benefit dramatically
from a top-view image with an overlay coordinate system. Added remark
that this is a right-handed coordinate system, otherwise y-axis would
not be uniquely specified.

Why do we go from output to input for the x-axis and not input to output
which would seem more natural (and would have shelf and slide on the
positive y side)?

git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@347 615d7b66-ede9-4446-987f-3b89ec2f13a0