machine setup -> prepare
Setup is used in multiple contexts. Then the integrator's manual and
refbox have not concept of setup of machines, there it's preparation.
Therefore, settle on preparation and change accordingly.

In the case of the RS, there was even a mistake, stating that additional
bases could only be filled in after setup/preparation. This has always
been different allowing for material to be filled in at any time.
Therefore, I took the liberty to change it accordingly.

git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@366 615d7b66-ede9-4446-987f-3b89ec2f13a0