Rephrase exploration
This is something we should have done last year already. We do accept
more than one report. It's the (partial) information we never update
once received. But to split position and orientation reporting, we
actually have to accept multiple reports.

Some other rephrasing in the hope to make things clearer.

git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@369 615d7b66-ede9-4446-987f-3b89ec2f13a0