Clarify on production plan
- Move SS products only for order with q_req >= 2 to production plan
section to keep this unified
- Change "orders of value 160" to "up to 10 orders. The previous also
only stated the same but less clear. I increased the number to 10 to
have some leeway in the schedule in the refbox. For example, we state
that there is one C1 with a no-cost ring color, but not all. So we might
actually want to have a second C1, for example the new standing C1
order. Clarify that one of the 10 orders is an overtime order.
- Remove some redundancies

git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/LLSF-RuleBook@373 615d7b66-ede9-4446-987f-3b89ec2f13a0