github: set CODEOWNERS to all individual TC members
If we just use a group name, the review process does not work properly.
More specifically, if one of the team members reviews a PR, the team is
not asked for a second review. As we want to get reviews from all team
members, this does not work for us.