penalize the dropping of workpieces on the field
Dropping workpieces on the field is an action equivalent to dropping
products or materials on the floor of a factory, which would require
additional clean-up. While it can be convenient for teams to just drop
workpieces, it should be penalized so that teams avoid such behavior
and instead use alternatives like shelves, storage stations, ring
payment slides or delivery stations.