rework scoring rules for late and wrong deliveries
Define late delivery penalties based on the delivery points awarded.
Also, increase the time window where the penalities build up. The
previous window of less than 15 seconds was so tight that it never
played a role.
Instead, take the delivery window length to determine the time frame,
where the penalty gradually increases.
Additionally, make the amount of deducted points dependant on the scored
delivery points when delivering late so that late deliveries actually
cause a significant drop in points for products with high complexity.

Finally, treat duplicate deliveries and wrong deliveries the same.
It makes no sense to award 1 point for duplicate deliveres as the amount
is so small that it never matters.